requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
redis==4.6.0
python-decouple==3.8
Pillow>=10.1.0
django-cors-headers==4.3.1
daphne==4.0.0
whitenoise==6.6.0
asgiref>=3.7.0